PXC-3050: Assertion when creating temporary table
https://jira.percona.com/browse/PXC-3050

Fixed wrong condition which prevented processing of writesets in case of dealing with temporary tables.

The problem was visible in the case when intermediate temporary InnoDB table was created when CREATE TEMPORARY TABLE was executed. That caused client_service to be in local mode, which allowed appending keys related to insertions to the intermediate temporary table.