PXC-3243 : BF aborting a stored procedure hangs
Issue
When running a SP with a transaction insde, BF-aborting the SP
can cause the system to hang.

Solution
This happens because the transaction is holding onto a lock,
but the BF-abort is aborting a statement, which may not be the
one holding the lock.
This change will force the BF-abort to propagate and abort and retry
the SP, not the statement.