Handle errors before or during storage engine initialization.
If an error happened before or during SE initialization, call
to `wsrep_deinit()` in `unireg_abort()` hung because the applier
thread was waiting for SE initialization and the provider side
cleanup could not be completed.

Changed SE_initialized boolean to enum carrying the result of
the SE initialization step. If there is an error and the server
will abort via `unireg_abort()` while the applier is waiting for
the SE initialization step to finish, the error is propagated via
SE_initialized to the applier thread, which will return an error
to the provider.

Due to the fix, the process exit will happen on Galera side,
leading to process exit code mismatch in tests

innodb-log-file-size-1
innodb-log-file-size
innodb_bug70867
innodb_page_size_func

Added tests into InnoDB suite galera.skip test list so that
they will be skipped when InnoDB suite is run with Galera loaded.