PXC-3381 : GTID functions broken in PXC 8.0.19
Issue
When calling WSREP_LAST_WRITTEN_GTID() and WSREP_LAST_SEEN_GTID()
from the MySQL client, they would return a hexadecimal string, instead
of a UUID:seqno.

This is a regression, see PS-7307. Previously (until 8.0.19), all UDFS
used CHARSET binary. MySQL clients would then display this binary data
as text.  Aflter WL#13038, the binary data is now displayed in hex
(--binary-as-hex is the default).  Thus our UUIDs are now displayed as
hex (instead of showing up as a UUID).

Solution
The solution is to change the functions to use the latin1 charset instead
of the bin charset. Thus the MySQL clients will interpret the data
as text and not convert the binary into hex.

Also added some basic functionality tests.