Restore wsrep recovery step in mysqld_safe
Due to a mistake during merging, mysqld_safe no longer performs
the recover step that is required to determine a proper starting
position for Galera.
This patch restores the wsrep recover step in mysqld_safe and adds a
MTR test that verifies that the recover step is performed.

Fixes codership/mysql-wsrep-bugs#833