PXC-2913: Fix MDL locks assertion when wsrep provider is unloaded
https://jira.percona.com/browse/PXC-2913

Problem:
This problem is visible only for debug builds and the case when wsrep
provider is not loaded (SET GLOBAL wsrep_provider = 'none')

1. Backup lock is taken
2. Transaction is started. This causes wsrep transaction object to be
created
3. Without involving any storage engine execute statement that causes
error. This causes wsrep transaction object transition to 'aborted'
state
4. Because wsrep provider is not loaded, wsrep_after_statement() is not
called and wsrep transaction id is not cleared
5. Execute next statement. wsrep_before_command() is executed and it
detects aborted transaction which causes MDL locks check (problematic
assert). However backup locks are allowed at this point.

Cause:
Wrong assert condition.

Solution:
Assert condition disallowing any MDL locks changed to the one
disallowing stmt/transactional locks. Explicit locks are allowed.

Note:
We still have the effect of the failure of the statement next to the one
causing wsrep transaction rollback. It fails with ER_LOCK_DEADLOCK. This
is caused by explained above detection of aborted transaction in
wsrep_before_command(). As this is the corner case and is the same as we
already have in release build, I decided not to change it. It would need
bigger logic rework, which seems not to be necessary for now.