Cherry-picked from 9a22fb944ed0fd68fe10950c99ee226055464ea4
PXC-3508 : SST with OpenSSL 1.1.1 and encrypt=2,3 fails with "E SSL_CTX_set_tmp_dh" from socat

Issue
Using OpenSSL fixed a bug with the way it used/generated dhparams.  This means that
certificates generated using an older version of OpenSSL would now fail if
used with OpenSSL 1.1.1.

Solution
Explicitly set the dhparam option with socat to workaround the use of the
old certs.