PXC-3565: PXC is more than 60% slower than PS
Issue: on longer running queries we can measure that PXC with wsrep
loaded is 60% slower than PXC without the wsrep library or PS.

Fix: After analyzing perf and callgrind results it is clear that
this is caused by the WSREP_ON macro, which is used two times for each
select. This macro had a strcmp call in it, which was executed two times
for each row.

To fix the performance issue this commit refactors the WSREP_ON macro:
instead of directly using strcmp it introduces the wsrep_provider_set
global boolean variable, which is recalculated when wsrep_provider
changes.