Implemented PXC-3566 (Merge PS-5.6.51)
https://jira.percona.com/browse/PXC-3566

***
Removed PS_REFRESH_VERSION as it is not used anywhere.

***
The test `innodb.innodb-autoinc` was failing because of the wrong merge
of upstream commit 1ccd472d

This caused the clients performing concurrent inserts with different
values of auto_increment_increment lead to "Duplicate key error" in one
of the sessions when wsrep is not loaded.