PXC-3536: Cluster data inconsistency after MDL conflict solved
Issue
DDL queries can be killed. If the query is killed before it is replicated,
this can lead to inconsistencies, because the primary node's DDL will have
been cancelled, but the replicas will have received and performed the
operation.

Solution
Do not allow threads/queries to be killed if the thread is in TOI.