codership/mysql-wsrep#383 mysqld --help --verbose tries to join the node to cluster
Skipping wsrep initialization is server is started with help option.

Added new test wsrep.help_verbose to verify that server is not trying to start replication if started with --help --verbose options.
This test is direct copy of the main help_verbose test, but just changed to use cluster configuration.