Fixed PXC-3596 (Node stuck in aborting SST) https://jira.percona.com/browse/PXC-3596
Problem:
SST xtrabackup process happens with an unlimited timeout on joiner side.
In case the donor crash or has a network issue it can trigger two issues
on joiner:
1) If the connection has already been establish to socat/nc, we will
relay on TCP layer to close the connection, this can vary depending on
the nature of the network issue and OS/Kernel configuration.
2) Socat will start to listen on the SST port and will block until a new
connection is established. If donor leaves the cluster after joiner has
started to listen but before the connection has been established, joiner
socat will hang forever.

Solution:
Implement a timeout function that will use the size of .sst directory
as metric to validate if we are suffering a stall on SST. In case of
data size has not moved since the last 5 seconds, we enter stall mode.
In stall mode we keep checking the size every second, until we reach
sst-timeout seconds (default to 120 seconds). When timeout is reached,
SST is aborted on joiner side.

This commit will be cherry picked at 8.0 branch