PXC-3559: Fix unstable galera_sr.GCF-1043B test
https://jira.percona.com/browse/PXC-3559

Problem:
Assertion is triggered when empty local transaction is BF aborted by
replicated transaction (originally reproduced by galera_sr.GCF-1043B
test ran in parallel)

Cause:
We removed check for TRX_FORCE_ROLLBACK_DISABLE from 'if' condition in
lock_make_trx_hit_list(), so now HP transaction can BF other transaction
even if the flag is set.
It is OK in general case if local transaction is not empty (so it is
replicated and certified). If it conflicts with HP transaction it will
not be certified (so will be discarded), and if it was certified, there
is no possibility for other transaction to BF it because that other
transaction should fail certification.
If local transaction is the empty transaction, it is not replicated
and certified as it does not generate any writestes. So it follows the
normal transaction path. But empty transaction can get some locks.
If at the same time there is replicated transaction attempting to get
the same locks, it can BF local transaction at any stage, even if local
transaction is commiting. (note that it can happen only for empty local
transaction)

Solution:
Allow trx_commit_for_mysql() to return DB_FORCED_ABORT.