Fixed PXC-3644 ([MTR] Tests fail during IST)
https://jira.percona.com/browse/PXC-3644

Problem:
This is a regression of 31323d496631ce93bb6d5c88ec71c05bdd74d3d7
SST / IST Script rans with set -u, which will abort the exection if
unbound variable is set. As part of previous fix, we were wrongly
testing if ${JOINER_SST_DIR} was set as part of cleanup.

Fix:
Adjusted test of ${JOINER_SST_DIR} to not halt the execution.