PXC-3555 - [MTR] Fix failing tests
Problem & Analysis
==================
The test `innodb.bootstrap` test fails to initialize due to the presence
of SSL files in data directory in case of a bootstrap failure.

This happens because of the `--pxc-encrypt-cluster-traffic` option (enabled
by default) which creates SSL files in the data directory, much before
the server bootstrap.

Fix
===
Disable `--pxc-encrypt-cluster-traffic` during the server
initialization.