PXC-3555 - [MTR] Fix failing tests
Problem
=======
The test `main.mysqltest` failed with execution timeout.

Analysis
========
PS' MTR doesn't allow names of the test and result files to contain
hash(#) character in it. But the same is an exception in PXC. As a
result, instead of erroring out when the file name had a '#' character
in it, mysqltest passed the case and was waiting for the user's input
from stdin, thereby causing the test to wait forever.

Fix
===
1. Modify the test to verify that '#' character is accepted in the test and
   the result files.
2. Modified the error message stating that the filename can have
   alpha-numeric characters,dash (-), underscore (_) or hash (#)
   characters in the filename.
3. Re-recorded the test with the new error message.