PXC-3619: Merge PS-8.0.23 (Resolved conflicts - sql/ & storage/)
***
Resolved conflicts in sql/ and storage/ directories.

***
Following up the changes in the Oracles's implementation of the
`Bug#31956584: RENAME MYSQL_PARSE() TO DISPATCH_SQL_COMMAND()`
(commit-id:e7141ab7), there were dependent code in PXC that had to be
updated because of this change.

- wsrep_mysql_parse() is changed to wsrep_dispatch_sql_command().
- Updated all occurences of `mysql_parse()` to `dispatch_sql_command()`.
- Fixed the build failures.

***
Following up the changes in the Oracles's implementation of the
'WL#14167: Connection alive in ServiceCommand.'
(https://dev.mysql.com/worklog/task/?id=14167)
(commit-id: 5a28a2fb).

Introduced `connection_alive()` callback functionality in
`Sql_service_context_base::sql_service_callbacks` in `wsrep_service.*`
files to make it possible to check if the connection is still alive when
using SQL session service.

***
Following up the changes in the Oracles's implementation of the
WL #14235 "Converting all old atomics code to C++11"
(https://dev.mysql.com/worklog/task/?id=14235)
(commit mysql/mysql-server@b5a4558),
`wsrep_killed_by_query` member of `trx_t` had to be changed to
`std::atomic`.

On further analysis, it was found that this member was a dead code and
was not being used anywhere. So, it has been removed from the sources.