Implemented PXC-3635 (Merge PS-5.7.34, Part 1 fixed build failures)
https://jira.percona.com/browse/PXC-3635

Following up the changes in the Oracles's implementation of the
(Bug#32354435: BACKPORT TO 5.7: BUG #32216281 REMOVE DBUG_OFF AND
               DBUG_ASSERT)
(commit 6889e5c0), where in MySQL dropped support for DBUG_ASSERT, there
were dependent code in PXC that had to be updated because of this
change.

Replaced DBUG_ASSERT with assert, and re-indented each region according
to appropriate C style in each source file.