PXC-3265: fix missing binlog writes with NBO
There was a remaining todo and an additional condition in the wsrep
binlog checks to fix a crash during development. The underlying issue
was already fixed, but the condition remained there, keeping the applier
thread from writing NBO statements to the binlog.

Also modified the applier thread to enable/disable binlogging based on
the log_slave_updates variable.