Bug#32552332 - CAN'T WRITE; DUPLICATE KEY IN TABLE '/TMP/#SQL...'
In the case of a query that uses temporary table for aggregation,
the group-by Item is used as the unique constraint of the temporary
table.

If the group-by Item value is already present, the row is updated.
Otherwise, a new row is inserted into the temporary table.

If the group-by Item is an Item with a result field(like Item_func)
or a reference Item, it gets evaluated twice. Once to check if the
result exists in the temporary table and if not, again while
constructing the row to insert.

If the group-by item is non-deterministic, the result value used to
check for existence would be different from the value with which an
insert is tried. This causes the insert to fail if the value already
exists in the table.

Fix:

If any of the group-by Items are non-deterministic, use the hash of
the group-by Items as the unique constraint of the temporary table as
the hash gets evaluated only once.