PXC-3739: FLUSH TABLES FOR EXPORT lock not released when session ends
https://jira.percona.com/browse/PXC-3739

Problem:
FLUSH TABLE(S) FOR EXPORT command pauses the provider and is not resumed
when session gets disconnected (or killed).

In PXC context, pausing the provider refers to entering the galera local
monitor, causing the local seqno to freeze and thereby blocking other
threads to commit. Similarily, resuming the cluster refers to leaving
the local monitor and allowing other waiting threads to proceed.

Solution:
Resume the provider if it was paused during FLUSH TABLE(S) FOR EXPORT
and when the session got disconnected (or killed).