Fixed PXC-3766 - Xtrabackup based SST always runs version-check procedure
https://jira.percona.com/browse/PXC-3766

Problem:
Half way thorught the life of xtrabackup/innobackupex, --version-check
functionality was added as feature that is turned on/off at compilation time.
This made SST script to validate if the server had support for --version-check,
if true we were adding --no-version-check parameter to skip it. By default, SST
was using innobackupex script. The script got deprecated and during the process
of migrating to xtrabackup binary the /tmp directory is no longer required as a
parameter for --help. Passing just /tmp will cause xtrabackup command to fail
and the check never resolving to true.

Fix:
Removed the /tmp path as parameter for discovering if xtrabackup was compiled
with version-check support.