PS-7989 Migrate innodb_encrypt_online_alter_logs to 8.0.27
https://jira.percona.com/browse/PS-7989

***

Problem:

`log_tmp_block_decrypt` uses the argument `offs` to work out the
initialization vector which was used in encryption. If the provided
`offs` is not the same as the one use in encryption the decrypted data
will be corrupted.

MySQL 8.0.27 has refactored the code so the `size` and `offs` are not
the same during write as during read so the current
`log_tmp_block_decrypt` is returning corrupted data.

Another upstream change, that affects these functions, is that the size
is not longer a multiple of MY_AES_BLOCK_SIZE. This causes we can't
neither use `my_aes_mode::ECB`.

The issue was discovered when running
`keyring_vault.innodb_online_alter_encryption`.

Solution:

Save the offsets used when the data was written and use those same
offset to read the data.

Testing:

- Affected tests:

- innodb.innodb_online_alter_encryption
- keyring_vault.innodb_online_alter_encryption
- keyring_vault.innodb_row_log_encryption
- innodb.innodb_row_log_encryption

- Run the `innodb` suite with this feature on:

```
--mysqld=--innodb-encrypt-online-alter-logs=ON \
--suite=innodb
```

If no encryption is set, AES is used.