PXC-3855: Assertion `!thd->in_active_multi_stmt_transaction() || thd->in_multi_stmt_transaction_mode()'
https://jira.percona.com/browse/PXC-3855

Problem
=======
Server hits the assertion

    !thd->in_active_multi_stmt_transaction() || thd->in_multi_stmt_transaction_mode()

when AUTOCOMMIT=ON and there is a binlog cache error when binlog is disabled
(i.e, when in binlog emulation mode)

Background
==========
1. When a server is in the binlog emulation mode and if there is an error
during the binlog cache write stage, we have a special handling in PXC which
makes the transaction to register with the binlog handlerton by calling
`trans_register_ha()` with binlog_hton, so that WSREP layer can can rollback
the transaction.

2. At the end of every statement, in `mysql_execute_command()` the server
   expects that

  - Session should not have SERVER_STATUS_IN_TRANS flag set.
  - If session has SERVER_STATUS_IN_TRANS flag set, then the session must be in
    multi statement transaction mode (either AUTOCOMMIT = OFF or inside a
    transaction started with BEGIN).

Analysis
========
When the server calls `trans_register_ha()` during binlog cache error, it sets
the `SERVER_STATUS_IN_TRANS` flag indicating that there was a beginning of a
multi-statement transaction, even when there was no multi statements involved
(because AUTOCOMMIT=ON).

This caused the assertion

    !thd->in_active_multi_stmt_transaction() || thd->in_multi_stmt_transaction_mode()

in `mysql_execute_command()` to fail, since we have SERVER_STATUS_IN_TRANS flag
set for the session and the session is not in a multi statement transaction mode
(since AUTOCOMMIT=ON), resulting in server abort.

Solution
========
In case of a binlog cache write error, register the transaction scope
binlog handler (call `trans_register_ha()` with all = true) only when
the server is in multi statement transaction mode, just like what is
being done in `binlog_start_trans_and_stmt()`.

In addition to the above fix, the test `wsrep.wsrep_set_var_extension`
has also been re-recorded.