PXC-3901: BF-BF conflict with the dict stats thread (#1597)
Issue: a few galera_nbo MTR tests fail with BF-BF conflict when ran with
the --mem options. These failures are all caused by the NBO applier
conflicting with the innodb dict stats thread.

Fix: as there is no way to detect if a THD is the dict stats thread,
this patch introduces a flag background threads can set to prevent BF-BF
abort caused by them.

Since the dict thread only holds this lock for a short time, and there
is no way to kill it anyway, this is the safest and easiest solution.