PXC-3923: ANALYZE TABLE crashes node when [super_]read_only is set
https://jira.percona.com/browse/PXC-3923

Problem:
When super_read_only or read_only is set, ANALYZE TABLE fails
(as expected) which triggers consistency voting mechanism and causes
node to leave the cluster

Cause:
In original flow the validation if admin action can be done in readonly
mode is done on transaction commit. As the statement is replicated as
TOI it is too late.

Solution:
Check if action can be done before starting TOI.