PXC-3951: Current computation of disk used with sst-idle-timeout could misidentify SST as idle
https://jira.percona.com/browse/PXC-3951

Problem
=======
Currently, disk usage computation of sst-idle-timeout is:

        du --max-depth=1 ${tmpsstdir} | tail -n 1 | awk '{print $1}'

This logic does not detect the storage size difference in bytes, thereby
causing the SST script to timeout and eventually terminate SST even when the
SST is in progress.

Solution
========
The script now passes `--bytes (-b)` and `--summarize (-s)` to `du` to
get difference of storage in bytes.