PXC-3961: PXC cross version replication fails between 5.6 and 5.7
https://jira.percona.com/browse/PXC-3961

Problem
-------
As part of the bugfix for
(BUG#32103192: REPLICA ACCEPTS GTID_LOG_EVENT WITH GNO=INT64_MAX)
(commit-id: 70cf38f9528fc5)
upstread added stricter checks to prevent the replica to apply a
Gtid_event that has a GTID with an invalid GNO number. This caused the
applier threads to mark the event with GNO<=0 and GNO>=INT64_MAX as
invalid.

However, in PXC, writesets are always replicated with GNO=0 because the actual
seqno/GTID is assigned from gcomm layer at a later stage (galera_pre_commit).
So, the assumption that "GNOs cannot be 0" does not hold true for galera
replication.

Solution
--------
Don't consider it as an error when GNO=0 when WSREP is loaded